Add new "UsesConstraint" entity.
The "UsesConstraint" connects an exported package to
the bundle declaring a "uses" constraint for that exported package.
This gives us the ability, in the case of multiple package providers,
to track back to which specific bundle is declaring a uses constraint
for the exported package.

Here's a cypher query to run against these new relationships:
MATCH (b)-[:EXPORTS]->(p)<-[:PACKAGE]-(uc:UsesConstraint)-[:USES]->(used)
RETURN b,p,uc,used LIMIT 10;